+++ src/cosmosimpy/CosmoSim.cpp
@@ -1,6 +1,7 @@
 /* (C) 2022: Hans Georg Schaathun <georg@schaathun.net> */
 
 #include "CosmoSim.h"
+
 #include <pybind11/pybind11.h>
 #include <opencv2/opencv.hpp>
