+++ src/simulator/Simulator.cpp
@@ -1,7 +1,7 @@
 /* (C) 2022: Hans Georg Schaathun <georg@schaathun.net> */
-#include "Simulator.h"
-#include "Window.h"
-#include <fstream>
+
+#include "cosmosim/Window.h"
+
 
 int main()
 {
